Rename gli module
