+++ tests/test_data_loading.py
@@ -3,8 +3,8 @@ import os
 import fnmatch
 import json
 import pytest
-import glb
-from glb.task import SUPPORTED_TASK_TYPES
+import gli
+from gli.task import SUPPORTED_TASK_TYPES
 from utils import find_datasets
 
 
@@ -33,7 +33,7 @@ def test_data_loading(dataset_name):
             print(os.path.splitext(file)[0])
             task_list.append(os.path.splitext(file)[0])
     try:
-        _ = glb.dataloading.get_glb_graph(dataset)
+        _ = gli.dataloading.get_glb_graph(dataset)
     except (AssertionError,
             AttributeError,
             ModuleNotFoundError,
@@ -44,7 +44,7 @@ def test_data_loading(dataset_name):
 
     for task in task_list:
         try:
-            _ = glb.dataloading.get_glb_task(dataset, task)
+            _ = gli.dataloading.get_glb_task(dataset, task)
         except (AssertionError,
                 AttributeError,
                 ModuleNotFoundError,
@@ -54,7 +54,7 @@ def test_data_loading(dataset_name):
             assert False
 
         try:
-            glb.dataloading.get_glb_dataset(dataset, task)
+            gli.dataloading.get_glb_dataset(dataset, task)
         except (AssertionError,
                 AttributeError,
                 ModuleNotFoundError,
